requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
fonttools==4.55.0
idna==3.10
iniconfig==2.0.0
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pluggy==1.5.0
pyparsing==3.2.0
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
tenacity==9.0.0
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
